pull out a couple utility functions
