+++ src/backend/commentsController.js
@@ -8,7 +8,6 @@ import { ThreadsService } from './threadsService'
 export const controller = (req, res) => {
   if (req.method === 'POST') {
     const { id } = req.query
-    console.log(id)
     const newComment = ThreadsService.addComment(id, JSON.parse(req.body))
     res.status(201).json(newComment)
   }
