+++ src/utils/sumValues.js
@@ -1,0 +1,3 @@
+export const sumValues = values => {
+  return values.reduce((sum, val) => val + sum, 0)
+}
